+++ jablib/src/main/java/org/jabref/logic/search/inmemory/InMemoryLuceneSearchBackend.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import org.jabref.logic.FilePreferences;
 import org.jabref.logic.search.SearchBackend;
-import org.jabref.logic.search.sqlbased.LinkedFilesIndexManager;
+import org.jabref.logic.search.LinkedFilesIndexManager;
 import org.jabref.logic.util.TaskExecutor;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
